cleanup-services/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.119.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-run==0.12.0
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
idna==3.11
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.3
pydantic-core==2.41.4
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
